+++ perf/documents.h
@@ -21,8 +21,8 @@
  * less than 4k, cause C99 says that's what we should do and YAJL likes streams */
 
 extern const char ** g_documents[];
-int num_docs(void);
-const char ** get_doc(int i);
-unsigned int doc_size(int i);
+unsigned int num_docs(void);
+const char ** get_doc(unsigned int i);
+size_t doc_size(unsigned int i);
 
 #endif
